[AND-163] Send feature flags fetch event only on network request
